+++ node/src/main/kotlin/net/corda/node/internal/AbstractNode.kt
@@ -32,11 +32,11 @@ import net.corda.core.transactions.SignedTransaction
 import net.corda.core.utilities.NetworkHostAndPort
 import net.corda.core.utilities.cert
 import net.corda.core.utilities.debug
-import net.corda.node.internal.classloading.CordappLoader
 import net.corda.node.internal.classloading.requireAnnotation
+import net.corda.node.internal.cordapp.CordappLoader
 import net.corda.node.services.NotaryChangeHandler
 import net.corda.node.services.NotifyTransactionHandler
-import net.corda.node.services.TransactionKeyHandler
+import net.corda.node.services.SwapIdentitiesHandler
 import net.corda.node.services.api.*
 import net.corda.node.services.config.NodeConfiguration
 import net.corda.node.services.config.configureWithDevSSLCertificate
@@ -69,6 +69,7 @@ import net.corda.node.services.vault.NodeVaultService
 import net.corda.node.services.vault.VaultSoftLockManager
 import net.corda.node.utilities.*
 import net.corda.node.utilities.AddOrRemove.ADD
+import net.corda.nodeapi.internal.serialization.DefaultWhitelist
 import org.apache.activemq.artemis.utils.ReusableLatch
 import org.slf4j.Logger
 import rx.Observable
@@ -135,9 +136,6 @@ abstract class AbstractNode(open val configuration: NodeConfiguration,
     lateinit var database: CordaPersistence
     protected var dbCloser: (() -> Any?)? = null
 
-    var isPreviousCheckpointsPresent = false
-        private set
-
     protected val _nodeReadyFuture = openFuture<Unit>()
     /** Completes once the node has successfully registered with the network map service
      * or has loaded network map data from local database */
@@ -149,18 +147,18 @@ abstract class AbstractNode(open val configuration: NodeConfiguration,
         CordaX500Name.build(cert.subject).copy(commonName = null)
     }
 
-    /** Fetch CordaPluginRegistry classes registered in META-INF/services/net.corda.core.node.CordaPluginRegistry files that exist in the classpath */
-    open val pluginRegistries: List<CordaPluginRegistry> by lazy {
-        ServiceLoader.load(CordaPluginRegistry::class.java).toList()
-    }
-
-    val cordappLoader: CordappLoader by lazy {
-        if (System.getProperty("net.corda.node.cordapp.scan.package") != null) {
+    private val cordappLoader: CordappLoader by lazy {
+        val scanPackage = System.getProperty("net.corda.node.cordapp.scan.package")
+        if (scanPackage != null) {
             check(configuration.devMode) { "Package scanning can only occur in dev mode" }
-            CordappLoader.createDevMode(System.getProperty("net.corda.node.cordapp.scan.package"))
+            CordappLoader.createDevMode(scanPackage)
         } else {
             CordappLoader.createDefault(configuration.baseDirectory)
         }
+    }
+
+    open val pluginRegistries: List<CordaPluginRegistry> by lazy {
+        cordappLoader.cordapps.flatMap { it.plugins } + DefaultWhitelist()
     }
 
     /** Set to true once [start] has been successfully called. */
@@ -205,17 +203,13 @@ abstract class AbstractNode(open val configuration: NodeConfiguration,
 
             makeVaultObservers()
 
-            checkpointStorage.forEach {
-                isPreviousCheckpointsPresent = true
-                false
-            }
             startMessagingService(rpcOps)
             installCoreFlows()
 
             installCordaServices()
             registerCordappFlows()
-            _services.rpcFlows += cordappLoader.findRPCFlows()
-            registerCustomSchemas(cordappLoader.findCustomSchemas())
+            _services.rpcFlows += cordappLoader.cordapps.flatMap { it.rpcFlows }
+            registerCustomSchemas(cordappLoader.cordapps.flatMap { it.customSchemas }.toSet())
 
             runOnStop += network::stop
         }
@@ -233,7 +227,7 @@ abstract class AbstractNode(open val configuration: NodeConfiguration,
     private class ServiceInstantiationException(cause: Throwable?) : Exception(cause)
 
     private fun installCordaServices() {
-        cordappLoader.findServices(info).forEach {
+        cordappLoader.cordapps.flatMap { it.services }.forEach {
             try {
                 installCordaService(it)
             } catch (e: NoSuchMethodException) {
@@ -275,7 +269,7 @@ abstract class AbstractNode(open val configuration: NodeConfiguration,
     }
 
     private fun registerCordappFlows() {
-        cordappLoader.findInitiatedFlows()
+        cordappLoader.cordapps.flatMap { it.initiatedFlows }
                 .forEach {
                     try {
                         registerInitiatedFlowInternal(it, track = false)
@@ -343,7 +337,7 @@ abstract class AbstractNode(open val configuration: NodeConfiguration,
         installCoreFlow(BroadcastTransactionFlow::class, ::NotifyTransactionHandler)
         installCoreFlow(NotaryChangeFlow::class, ::NotaryChangeHandler)
         installCoreFlow(ContractUpgradeFlow.Initiator::class, ::Acceptor)
-        installCoreFlow(TransactionKeyFlow::class, ::TransactionKeyHandler)
+        installCoreFlow(SwapIdentitiesFlow::class, ::SwapIdentitiesHandler)
     }
 
     /**
@@ -486,8 +480,8 @@ abstract class AbstractNode(open val configuration: NodeConfiguration,
         val address: SingleMessageRecipient = networkMapAddress ?:
                 network.getAddressOfParty(PartyInfo.Node(info)) as SingleMessageRecipient
         // Register for updates, even if we're the one running the network map.
-        return sendNetworkMapRegistration(address).flatMap { response: RegistrationResponse ->
-            check(response.error == null) { "Unable to register with the network map service: ${response.error}" }
+        return sendNetworkMapRegistration(address).flatMap { (error) ->
+            check(error == null) { "Unable to register with the network map service: $error" }
             // The future returned addMapService will complete on the same executor as sendNetworkMapRegistration, namely the one used by net
             services.networkMapCache.addMapService(network, address, true, null)
         }