+++ node/src/test/kotlin/net/corda/node/services/config/FullNodeConfigurationTest.kt
@@ -14,7 +14,7 @@ class FullNodeConfigurationTest {
     @Test
     fun `Artemis special characters not permitted in RPC usernames`() {
         val testConfiguration = FullNodeConfiguration(
-                basedir = Paths.get("."),
+                baseDirectory = Paths.get("."),
                 myLegalName = ALICE.name,
                 networkMapService = null,
                 emailAddress = "",