+++ samples/irs-demo/src/test/kotlin/net/corda/irs/Main.kt
@@ -2,11 +2,10 @@ package net.corda.irs
 
 import net.corda.core.node.services.ServiceInfo
 import net.corda.core.utilities.getOrThrow
+import net.corda.node.services.transactions.SimpleNotaryService
 import net.corda.testing.DUMMY_BANK_A
 import net.corda.testing.DUMMY_BANK_B
 import net.corda.testing.DUMMY_NOTARY
-import net.corda.irs.api.NodeInterestRates
-import net.corda.node.services.transactions.SimpleNotaryService
 import net.corda.testing.driver.driver
 
 /**
@@ -15,12 +14,13 @@ import net.corda.testing.driver.driver
  */
 fun main(args: Array<String>) {
     driver(dsl = {
-        val controllerFuture = startNode(
-                providedName = DUMMY_NOTARY.name,
-                advertisedServices = setOf(ServiceInfo(SimpleNotaryService.type), ServiceInfo(NodeInterestRates.Oracle.type)))
-        val nodeAFuture = startNode(providedName = DUMMY_BANK_A.name)
-        val nodeBFuture = startNode(providedName = DUMMY_BANK_B.name)
-        val (controller, nodeA, nodeB) = listOf(controllerFuture, nodeAFuture, nodeBFuture).map { it.getOrThrow() }
+        val (controller, nodeA, nodeB) = listOf(
+                startNode(
+                        providedName = DUMMY_NOTARY.name,
+                        advertisedServices = setOf(ServiceInfo(SimpleNotaryService.type))),
+                startNode(providedName = DUMMY_BANK_A.name),
+                startNode(providedName = DUMMY_BANK_B.name))
+                .map { it.getOrThrow() }
 
         startWebserver(controller)
         startWebserver(nodeA)